FAGD but without the idiot proof bloat
Finally! after popular demand by the eediot that helped me with the HTML crap, I decided to make a version for people who actually know how to Linux and  don't need someone to hold their hand on something as simple as getting the right packages and starting up a service.